pyproject.toml
[project]
name = "grimorium"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Malcom Godlike", email = "ivmno1special@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.32.0",
    "scikit-learn>=1.7.1",
    "pyyaml>=6.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.1.1",
    "google-adk[a2a]>=1.13.0",
    "chromadb>=1.4.1",
]

[build-system]
requires = ["uv_build>=0.8.8,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "black>=26.1.0",
    "isort>=7.0.0",
]

